+++ academy-backend/src/main/java/uz/jurayev/academy/repository/CountryRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface CountryRepository extends JpaRepository<Country, Long> {
     boolean existsByCountryName(String countryName);
 
-    Optional<Country> findById(Integer id);
+    Optional<Country> findById(Long id);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
