+++ academy-backend/src/main/java/uz/jurayev/academy/service/impl/AddressServiceImpl.java
@@ -50,13 +50,13 @@ public class AddressServiceImpl implements AddressService {
     }
 
     @Override
-    public Address getOne(Integer id) {
+    public Address getOne(Long id) {
         Optional<Address> byId = addressRepository.findById(id);
         return byId.orElse(null);
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Long id) {
         try {
             addressRepository.deleteById(id);
             return true;
@@ -66,7 +66,7 @@ public class AddressServiceImpl implements AddressService {
     }
 
     @Override
-    public Address edit(Integer id, AddressDTO addressDTO) {
+    public Address edit(Long id, AddressDTO addressDTO) {
         Optional<District> optionalDistrict = districtRepository.findById(addressDTO.getDistrictId());
         Optional<Address> byId = addressRepository.findById(id);
